Refactor: Move checkTradeAmount and checkTakersTradePrice to DepositTxValidation
