+++ core/src/main/java/bisq/core/trade/validation/TradeValidation.java
@@ -22,10 +22,7 @@ import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.Restrictions;
 import bisq.core.btc.wallet.TradeWalletService;
 import bisq.core.dao.burningman.DelayedPayoutTxReceiverService;
-import bisq.core.exceptions.TradePriceOutOfToleranceException;
 import bisq.core.offer.Offer;
-import bisq.core.offer.OfferValidation;
-import bisq.core.offer.bisq_v1.MarketPriceNotAvailableException;
 import bisq.core.provider.fee.FeeService;
 import bisq.core.provider.price.PriceFeedService;
 import bisq.core.support.dispute.mediation.mediator.Mediator;
@@ -89,34 +86,6 @@ public class TradeValidation {
     /* --------------------------------------------------------------------- */
     // Offer
     /* --------------------------------------------------------------------- */
-
-    public static Coin checkTradeAmount(Coin tradeAmount, Coin offerMinAmount, Coin offerMaxAmount) {
-        checkIsPositive(tradeAmount, "tradeAmount");
-        checkIsPositive(offerMinAmount, "offerMinAmount");
-        checkIsPositive(offerMaxAmount, "offerMaxAmount");
-
-        checkArgument(!tradeAmount.isLessThan(offerMinAmount),
-                "Trade amount must not be less than minimum offer amount. tradeAmount=%s, offerMinAmount=%s",
-                tradeAmount.toFriendlyString(), offerMinAmount.toFriendlyString());
-        checkArgument(!tradeAmount.isGreaterThan(offerMaxAmount),
-                "Trade amount must not be higher than maximum offer amount. tradeAmount=%s, offerMaxAmount=%s",
-                tradeAmount.toFriendlyString(), offerMaxAmount.toFriendlyString());
-        return tradeAmount;
-    }
-
-    public static long checkTakersTradePrice(long takersTradePrice,
-                                             PriceFeedService priceFeedService,
-                                             Offer offer) {
-        try {
-            offer.verifyTakersTradePrice(takersTradePrice);
-            // We allow 50% tolerance to the max allowed price percentage to avoid failing trades in
-            // high volatility environments
-            OfferValidation.verifyPriceInBounds(priceFeedService, offer, MAX_TRADE_PRICE_DEVIATION);
-            return takersTradePrice;
-        } catch (TradePriceOutOfToleranceException | MarketPriceNotAvailableException e) {
-            throw new RuntimeException(e);
-        }
-    }
 
 
     /* --------------------------------------------------------------------- */
@@ -582,7 +551,7 @@ public class TradeValidation {
         checkNotNull(delayedPayoutTxReceiverService, "delayedPayoutTxReceiverService must not be null");
         checkNotNull(feeService, "feeService must not be null");
 
-        Coin tradeAmount = checkTradeAmount(request.getTradeAmountAsCoin(), offer.getMinAmount(), offer.getAmount());
+        Coin tradeAmount = DepositTxValidation.checkTradeAmount(request.getTradeAmountAsCoin(), offer.getMinAmount(), offer.getAmount());
         Coin tradeTxFee = checkTradeTxFeeIsInTolerance(request.getTxFeeAsCoin(), feeService);
         checkTakersRawTransactionInputs(request.getRawTransactionInputs(),
                 btcWalletService,
@@ -602,7 +571,7 @@ public class TradeValidation {
         checkPeersDate(request.getCurrentDate());
         NodeAddress mediatorNodeAddress = request.getMediatorNodeAddress();
         getCheckedMediatorPubKeyRing(mediatorNodeAddress, user);
-        checkTakersTradePrice(request.getTradePrice(), priceFeedService, offer);
+        DepositTxValidation.checkTakersTradePrice(request.getTradePrice(), priceFeedService, offer);
         checkTakerFee(request.getTakerFeeAsCoin(), request.isCurrencyForTakerFeeBtc(), tradeAmount);
         return request;
     }
